Native source codes to compare performance
Three new source files have been added.
The three of them implement a simple vector addition in three
different languages: OpenACC, CUDA and OpenCL.
The idea is to compare the times in the critical region and the total
execution time of the programs to see how much overhead the OpenACC
compiler adds when compared with using the native implementatons.